+++ src/main/java/com/ybe/tr1ll1on/domain/product/controller/ProductController.java
@@ -5,16 +5,13 @@ import com.ybe.tr1ll1on.domain.product.dto.response.AccommodationDetailResponse;
 import com.ybe.tr1ll1on.domain.product.dto.request.AccommodationRequest;
 import com.ybe.tr1ll1on.domain.product.dto.response.ProductResponse;
 import com.ybe.tr1ll1on.domain.product.service.ProductService;
-import com.ybe.tr1ll1on.global.date.utill.DateUtil;
-import jakarta.validation.Valid;
+import com.ybe.tr1ll1on.global.date.util.DateUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import java.time.LocalDate;
 
 @RestController
 @RequiredArgsConstructor
@@ -26,29 +23,60 @@ public class ProductController {
     @PostMapping("/{accommodation_id}")
     public ResponseEntity<AccommodationDetailResponse> getAccommodationDetail(
         @PathVariable("accommodation_id") final Long accommodationId,
-        @Valid @RequestBody final AccommodationRequest accommodationRequest
+        @RequestParam(required = false) LocalDate checkIn,
+        @RequestParam(required = false) LocalDate checkOut,
+        @RequestParam(required = false) Integer personNumber
     ) {
+        AccommodationRequest request = new AccommodationRequest();
+        if (checkIn != null) {
+            request.setCheckIn(checkIn);
+        }
+
+        if (checkOut != null) {
+            request.setCheckOut(checkOut);
+        }
+
+        if (personNumber != null) {
+            request.setPersonNumber(personNumber);
+        }
 //        TODO checkIn 과 checkOut은 같은 날짜이면 안됨!!
         DateUtil.isValidCheckInBetweenCheckOut(
-                accommodationRequest.getCheckIn(), accommodationRequest.getCheckOut()
+                request.getCheckIn(), request.getCheckOut()
         );
+
         return ResponseEntity.ok(
-                productService.getAccommodationDetail(accommodationId, accommodationRequest)
+                productService.getAccommodationDetail(accommodationId, request)
         );
     }
 
-    @PostMapping("/{accommodation_id}/{product_id}")
+    @GetMapping("/{accommodation_id}/{product_id}")
     public ResponseEntity<ProductResponse> getProductDetail(
             @PathVariable("accommodation_id") final Long accommodationId,
             @PathVariable("product_id") final Long product_id,
-            @Valid @RequestBody final AccommodationRequest accommodationRequest
+            @RequestParam(required = false) LocalDate checkIn,
+            @RequestParam(required = false) LocalDate checkOut,
+            @RequestParam(required = false) Integer personNumber
     ) {
+        AccommodationRequest request = new AccommodationRequest();
+        if (checkIn != null) {
+            request.setCheckIn(checkIn);
+        }
+
+        if (checkOut != null) {
+            request.setCheckOut(checkOut);
+        }
+
+        if (personNumber != null) {
+            request.setPersonNumber(personNumber);
+        }
+
         //TODO checkIn 과 checkOut은 같은 날짜이면 안됨!!
         DateUtil.isValidCheckInBetweenCheckOut(
-                accommodationRequest.getCheckIn(), accommodationRequest.getCheckOut()
+                request.getCheckIn(), request.getCheckOut()
         );
+
         return ResponseEntity.ok(
-                productService.getProduct(accommodationId, product_id, accommodationRequest)
+                productService.getProduct(accommodationId, product_id, request)
         );
     }
 }
